+++ src/Plugin.php
@@ -20,6 +20,8 @@ use craft\queue\Queue;
 use ostark\AsyncQueue\Exceptions\LogicException;
 use ostark\AsyncQueue\Exceptions\PhpExecutableNotFound;
 use ostark\AsyncQueue\Exceptions\RuntimeException;
+use ostark\AsyncQueue\Handlers\BackgroundQueueHandler;
+use ostark\AsyncQueue\Handlers\ProcessPoolCleanupHandler;
 use ostark\AsyncQueue\TestUtility\Utility;
 use yii\base\ActionEvent;
 use yii\base\Event;
@@ -53,7 +55,7 @@ class Plugin extends BasePlugin
 
         // Register plugin components
         $this->setComponents([
-            'async_handler' => QueueHandler::class,
+            'async_process' => BackgroundProcess::class,
             'async_pool'    => ProcessPool::class,
         ]);
 
@@ -61,8 +63,10 @@ class Plugin extends BasePlugin
         Craft::$container->set(CacheInterface::class, Craft::$app->getCache());
 
         // Register event handlers
-        PushEvent::on(Queue::class, Queue::EVENT_AFTER_PUSH, [$this, 'runQueueInBackground']);
-        Event::on(Command::class, Command::EVENT_AFTER_ACTION, [$this, 'freeProcessPool']);
+        PushEvent::on(Queue::class, Queue::EVENT_AFTER_PUSH, new BackgroundQueueHandler($this));
+        Event::on(Command::class, Command::EVENT_AFTER_ACTION, new ProcessPoolCleanupHandler($this));
+
+        // Register CP Utility
         Utility::setup($this);
 
     }
@@ -72,11 +76,11 @@ class Plugin extends BasePlugin
     // =========================================================================
 
     /**
-     * @return \ostark\AsyncQueue\QueueHandler
+     * @return \ostark\AsyncQueue\BackgroundProcess
      */
-    public function getHandler(): QueueHandler
+    public function getProcess(): BackgroundProcess
     {
-        return $this->get('async_handler');
+        return $this->get('async_process');
     }
 
     /**
@@ -85,60 +89,6 @@ class Plugin extends BasePlugin
     public function getPool(): ProcessPool
     {
         return $this->get('async_pool');
-    }
-
-    // EventHandlers
-    // =========================================================================
-
-    /**
-     * @param \yii\queue\PushEvent $event
-     */
-    public function runQueueInBackground(PushEvent $event)
-    {
-        // Disable frontend queue runner
-        Craft::$app->getConfig()->getGeneral()->runQueueAutomatically = false;
-
-        $context = ($event->job instanceof JobInterface)
-            ? $event->job->getDescription()
-            : 'Not instanceof craft\queue\JobInterface';
-
-        // Run queue in the background
-        if ($this->getPool()->canIUse($context)) {
-            try {
-                $this->getHandler()->startBackgroundProcess();
-                $this->getPool()->increment($context);
-                $handled = true;
-            } catch (PhpExecutableNotFound $e) {
-                Craft::debug(
-                    'QueueHandler::startBackgroundProcess() (PhpExecutableNotFound)',
-                    'async-queue'
-                );
-            } catch (RuntimeException | LogicException $e) {
-                Craft::debug(
-                    Craft::t(
-                        'async-queue',
-                        'QueueHandler::startBackgroundProcess() (Job status: {status}. Exit code: {code})', [
-                            'status' => $e->getProcess()->getStatus(),
-                            'code'   => $e->getProcess()->getExitCodeText()
-                        ]
-                    ),
-                    'async-queue'
-                );
-            }
-        }
-
-        // Log what's going on
-        $this->logPushEvent($event, $handled ?? false);
-    }
-
-    /**
-     * @param \yii\base\ActionEvent $event
-     */
-    public function freeProcessPool(ActionEvent $event)
-    {
-        if ('run' === $event->action->id) {
-            $this->getPool()->decrement(Command::class . '::run() ' . Command::EVENT_AFTER_ACTION);
-        }
     }
 
 
@@ -152,26 +102,4 @@ class Plugin extends BasePlugin
         return new Settings();
     }
 
-    /**
-     * @param \yii\queue\PushEvent $event
-     * @param bool                 $handled
-     */
-    protected function logPushEvent(PushEvent $event, $handled = false)
-    {
-        if (!YII_DEBUG) {
-            return;
-        }
-        if ($event->job instanceof BaseJob) {
-            Craft::debug(
-                Craft::t(
-                    'async-queue',
-                    'New PushEvent for {job} job - ({handled})', [
-                        'job'     => $event->job->getDescription(),
-                        'handled' => $handled ? 'handled' : 'skipped'
-                    ]
-                ),
-                'async-queue'
-            );
-        }
-    }
 }