+++ src/TestUtility/Controller.php
@@ -1,6 +1,7 @@
 <?php namespace ostark\AsyncQueue\TestUtility;
 
 use craft\web\Controller as BaseController;
+use ostark\AsyncQueue\Events\TestJobEvent;
 use yii\web\Response;
 
 /**
@@ -17,14 +18,15 @@ class Controller extends BaseController
 
     public function actionRun(): Response
     {
+        // Push the job multiple times to the queue
         foreach (range(1, 10) as $counter) {
             \Craft::$app->getQueue()->push(new TestJob([
-                'counter' => $counter
+                'counter' => $counter,
             ]));
         }
+
+        // Redirect back
         return $this->redirectToPostedUrl();
     }
 }
-
-