+++ src/TestUtility/TestJob.php
@@ -21,6 +21,7 @@ class TestJob extends BaseJob
         sleep(10);
     }
 
+
     public function defaultDescription()
     {
         return 'Async Queue Test Job ' . $this->counter;